Integrate side menu and header with map and right sidebar
