+++ package.json
@@ -49,6 +49,7 @@
     "chart.js": "^2.6.0",
     "jquery-ui-dist": "^1.12.1",
     "leaflet": "^1.0.3",
+    "material-ui": "^0.20.0",
     "materialize-css": "^0.100.1",
     "node-sass": "^4.5.3",
     "react-breadcrumbs": "^1.5.2",
