+++ src/js/routes.js
@@ -17,14 +17,13 @@ import Todo from './views/utils/todo';
 export default (
   <Router history={hashHistory}>
     <Route path="/" component={Full}>
-      <IndexRoute component={DeviceDashboard} />
-      <Route path="dashboard" name="Home" component={DeviceDashboard} />
+      <IndexRoute component={Devices} />
       <Route name="Device manager" >
-        <Route path="deviceManager" name="Device manager" component={DeviceDashboard} />
+        <Route path="deviceManager" name="Device manager" component={Devices} />
         <Route path="device" name="Devices">
-          <IndexRoute component={DeviceDashboard} />
+          <IndexRoute component={Devices} />
           <Route path="list" name="Device list" component={Devices} />
-          <Route path="stats" name="Device Dashboard" component={DeviceDashboard} />
+          <Route path="stats" name="Device Dashboard" component={Devices} />
           <Route path="new" name="" component={NewDevice} />
           <Route path="id/:device/detail" name="Device detail" component={ViewDevice} />
           <Route path="id/:device/edit" name="Device edit" component={NewDevice} />
