+++ src/sass/_device.scss
@@ -12,7 +12,7 @@
 
   position: relative;
   width: 100%;
-  height: $page-header-size;
+  //height: $page-header-size;
   animation-name: height-opacity-0-to-100;
   animation-duration: 1.5s;
   animation-iteration-count: 1;
@@ -21,9 +21,13 @@
   .main-title{
     font-family: 'Raleway', sans-serif;
     padding-left:4em;
+    position:relative;
+    top: 10px;
   }
 
   .header-info{
+    position:relative;
+    top: 10px;
     .title
     {
       font-family: 'Open Sans', sans-serif;
@@ -42,13 +46,321 @@
 
   .btn-flat{
     position: absolute;
-    right: 120px;
+    top:15px;
+    right: 130px;
   }
 }
 
 .deviceCanvas{
   padding-top: $page-header-size + 5px;
   top:-50px;
+}
+
+
+.leaflet-container {
+  height:100%;
+}
+
+.mapOptions{
+  position: absolute;
+  left: 60px !important;
+  z-index: 450;
+  top: 13px;
+  cursor: pointer;
+
+  .mapView{
+    background-color: white;
+    color: black;
+    font-size: 1em;
+    width:50px;
+    height:25px;
+    padding: 5px;
+    border-right: 1px solid #BDBDBD;
+  }
+
+  .satelliteView{
+    background-color: white;
+    color: black;
+    font-size: 1em;
+    width:65px;
+    height:25px;
+    position: relative;
+    top: -25px;
+    left: 50px;
+    padding: 5px;
+  }
+}
+
+.deviceMapCanvas
+{
+  height: 100%;
+  overflow: auto;
+  display: inline-flex;
+  >div
+  {
+    height: auto;
+    display: inline-block;
+  }
+
+  .row
+  {
+    margin: 0;
+    width: 100%;
+  }
+}
+
+.deviceMapCanvas-map
+{
+  height: 100%;
+}
+
+
+.leaflet-container {
+  height: 100%;
+  width: 100%;
+}
+
+
+.device-painel
+{
+  position: absolute;
+  z-index: 1000;
+  background: rgba(255, 255, 255, 0.98);
+  right: 0;
+  display: inline-block;
+  top:0;
+  width: 400px;
+  box-shadow: 10px 0px 50px #000000;
+}
+
+.device-painel-body
+{
+    overflow: auto;
+    height: calc(100% - 35px);
+}
+.list-of-devices{
+    float: left;
+    height: 100%;
+    width: 100%;
+   position: relative;
+   animation: fadein 200ms;
+ }
+
+.device-list{
+  width: 100%;
+  height: 100%;
+  float: right;
+  margin: 0;
+
+  .main-title{
+    font-family: 'Open Sans', sans-serif;
+    font-weight: bold;
+    font-size: 15px;
+    text-align: center;
+    letter-spacing: 0.39px;
+    text-transform: uppercase;
+    padding: 20px 10px 6px 10px;
+    color: #b3b3b3;
+  }
+
+  .info-header{
+    padding: 10px 10px 0 10px;
+    border-bottom: 1px solid #d9d9d9;
+    width: 100%;
+    float: left;
+    font-weight: bold;
+
+    .title{
+      font-family:"Roboto", sans-serif;
+      color: #A2A2A2;
+    }
+
+    .subtitle{
+      font-family:"Roboto", sans-serif;
+      color: #A2A2A2;
+    }
+
+    .device-list-actions{
+      position: relative;
+      a{
+        color: #A2A2A2;
+      }
+      .action-hide{
+        font-family:"Roboto", sans-serif;
+        color: #A2A2A2;
+        font-size: 0.8em;
+      }
+
+      .action-show{
+        font-family:"Roboto", sans-serif;
+        color: #A2A2A2;
+        font-size: 0.8em;
+      }
+    }
+  }
+
+  .deviceCanvas{
+    width: 100% !important;
+
+    .background-info{
+      float: left;
+      width: 100%;
+      bottom: 0;
+      right: 0;
+      cursor: pointer;
+      background-color: #FAFAFA;
+      a{
+        font-family: 'Open Sans', sans-serif;
+        color: #DADADA;
+        font-weight: bold;
+        font-size: 0.8em;
+      }
+    }
+
+    .row{
+      margin:0;
+      .lst-entry-title{
+        padding: 0 10px;
+        height: 60px;
+        transition: linear 100ms;
+        border-bottom: 1px solid #e2e2e2;
+        cursor:pointer;
+        .img
+        {
+          width: 58px;
+          padding: 10px;
+          margin: 0 10px;
+          img
+          {
+              width: 100%;
+          }
+        }
+        .waves-effect{
+          overflow: visible !important;
+          width: 90%;
+        }
+        &:hover{
+          background-color: rgba(0, 0, 0, 0.1);
+          transition: linear 100ms;
+        }
+        .user-label{
+          text-transform: uppercase;
+          padding-top: 6px;
+          font-family: 'Open Sans', sans-serif;
+          color: #4F4F4F;
+          font-weight: normal;
+        }
+        .label{
+          font-family: 'Open Sans', sans-serif;
+          color: #4F4F4F;
+          font-weight: bold;
+
+        }
+        #device-view{
+          position: relative;
+          text-align: center;
+          top: -28px;
+          padding: 20px;
+          float: right;
+          i{
+            color: #4f4f4f;
+          }
+        }
+      }
+    }
+  }
+}
+
+.div-btn-side-painel{
+  position: absolute;
+  right: 20px;
+  top: 20px;
+  z-index: 100110;
+
+  &.push-left{
+    right: 386px;
+  }
+}
+
+.toggle-icon
+{
+  width: 17px;
+  height: auto;
+  display: inline-block;
+  margin: 0 3px;
+  float:left;
+
+   i
+   {
+       font-size: 21px;
+   }
+
+
+  img
+  {
+    width: 100%;
+    height: auto;
+  }
+}
+
+.toggle-map
+{
+  width: 56px;
+  float: left;
+  display: inline-block;
+}
+
+
+.box-sh
+{
+  padding: 20px;
+  width: auto;
+  position: relative;
+  height: 100%;
+  display: inline-block;
+  float: right;
+
+}
+
+.box-sh-item.active {
+  border-bottom: 5px solid #425A8D;
+  color: #425A8D;
+}
+
+
+.box-sh-item {
+  padding: 16px 10px 10px 10px;
+  width: auto;
+  border-bottom: 5px solid #BDBDBD;
+  color: #828282;
+  height: 100%;
+  display: inline-block;
+  float: left;
+  margin: 0 10px;
+
+  .icon
+  {
+    width: 26px;
+    display: inline-block;
+    float: left;
+    height: 100%;
+    position: relative;
+    padding-top: 3px;
+
+    img {
+      width: 100%;
+    }
+  }
+
+
+  .text
+  {
+    display: inline-block;
+    padding-left: 10px;
+    float: left;
+    text-transform: uppercase;
+    margin: 7px;
+  }
 }
 
 @media screen and (max-width: 600px){
